+++ lib/action_cable/channel/base.rb
@@ -190,7 +190,7 @@ module ActionCable
       # confirms or rejects the subscription.
       def subscribe_to_channel
         run_callbacks :subscribe do
-          subscribed
+          subscribed unless subscription_rejected?
         end
 
         reject_subscription if subscription_rejected?